pyproject.toml
[project]
name = "windows-gui"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "pillow==8.1.0",
    "pyautogui>=0.9.54",
    "opencv-python>=4.10.0.84",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/windows_gui"]
